services/am-email-extractor/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0
PyPDF2==3.0.1
openpyxl==3.1.2
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.1
kafka-python==2.0.2
pdfplumber==0.10.3
pandas==2.1.3
